remove pages_old + pages files
