+++ examples/pages/frontend/src/admin.rs
@@ -1,6 +1,3 @@
-use zoon::*;
-use crate::app;
-
 mod report;
 
 blocks!{
