+++ examples/pages/frontend/src/admin/report.rs
