--- examples/pages_old/Cargo.toml
@@ -1,6 +1,0 @@
-[workspace]
-
-members = [
-    "frontend",
-    "backend",
-]
