--- examples/pages_old/README.md
@@ -1,2 +1,0 @@
-# Pages
-> MoonZoon example
