--- examples/pages_old/backend/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-version = "0.1.0"
-name = "backend"
-repository = "https://github.com/MoonZoon/MoonZoon"
-authors = ["Martin Kavík <martin@kavik.cz>"]
-description = "Pages - MoonZoon example"
-categories = ["MoonZoon", "example"]
-license = "MIT"
-readme = "../README.md"
-edition = "2018"
-
-[dependencies]
-moon = { path = "../../../moon" }
-
-[profile.release]
-lto = true
-codegen-units = 1
