--- examples/pages_old/backend/src/main.rs
@@ -1,15 +1,0 @@
-use moon::*;
-
-async fn init() {}
-
-async fn frontend() -> Frontend {
-    Frontend::new()
-        .title("Counter example")
-        .css(include_str!("../../style.css"))
-}
-
-async fn up_msg_handler(req: UpMsgRequest) {}
-
-fn main() {
-    start!(init, frontend, up_msg_handler);
-}
