--- examples/pages_old/frontend/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-version = "0.1.0"
-name = "frontend"
-repository = "https://github.com/MoonZoon/MoonZoon"
-authors = ["Martin Kavík <martin@kavik.cz>"]
-description = "Pages - MoonZoon example"
-categories = ["MoonZoon", "example"]
-license = "MIT"
-readme = "../README.md"
-edition = "2018"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dev-dependencies]
-wasm-bindgen-test = "0.3.19"
-
-[dependencies]
-zoon = { path = "../../../zoon" }
-
-[profile.release]
-lto = true
-opt-level = 'z'
-codegen-units = 1
-
-[package.metadata.wasm-pack.profile.release]
-wasm-opt = ['-Os']
-
