--- examples/pages_old/frontend/src/lib.rs
@@ -1,10 +1,0 @@
-#![no_std]
-
-use zoon::*;
-
-mod app;
-
-#[wasm_bindgen(start)]
-pub fn start() {
-    start!(app)
-}
