--- examples/pages_old/shared/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-version = "0.1.0"
-name = "shared"
-repository = "https://github.com/MoonZoon/MoonZoon"
-authors = ["Martin Kavík <martin@kavik.cz>"]
-description = "Pages - MoonZoon example"
-categories = ["MoonZoon", "example"]
-license = "MIT"
-readme = "../README.md"
-edition = "2018"
