--- examples/pages_old/shared/src/lib.rs
@@ -1,1 +1,0 @@
-
