--- examples/pages_old/style.css
@@ -1,5 +1,0 @@
-html, body {
-    height: 100%;
-    padding: 0;
-    margin: 0;
-}
